refactor(router): extract routes/mesh.rs (q1 wave 2/5)
Move mesh/relay/blocklist handlers + mesh_routes() builder into
routes/mesh.rs:

  - agt_relay_proxy, relay_websocket_bridge
  - agt_registry_proxy
  - lookup_parent_amid (exposed as pub(super) — called from handoff
    in mod.rs at 3 sites; stays reachable without behavioral change)
  - blocklist_status, blocklist_check

Public API unchanged: routes::mesh_routes() re-exported via pub use.

mod.rs also drops the now-unused axum::extract::ws::* imports.

Proof of behavioral equivalence:
- 178/178 router tests green
- cargo clippy clean
- item-manifest drift-check: ZERO drift — every function body
  byte-identical, every data item preserved.
